new file:   opensourceleg/encoder.py 	new file:   tests/test_example/test_encoder.py 	modified:   tests/test_example/test_hello.py
Started work on TMotor integration with Encoder abstract class and AS5048A_Encoder implementation

Encoder defines abstract interfaces such as
	open()
	close()
	encoder_position()
	encoder_velocity()
	[...]
These are implemented by subclassing the hardware implementations. The goals is to
allow OSL or Joint to use the Encoder interface without caring about the specific implementations.

AS5048A_Encoder is a rough porting of the existing code in the Straing guage development-branch.
It has been roughly tested with a hardware setup and is working.